api/requirements.txt
azure-core==1.35.1
azure-cosmos==4.9.0
azure-functions==1.23.0
azure-identity==1.25.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==46.0.1
Flask==2.3.3
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msal==1.33.0
msal-extensions==1.3.1
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.15
packaging==24.2
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pywatchman==2.0.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1